[issue-221] Move the survey files to new directory
