+++ cocoon/frontend/src/userAuth/survey/survey.js
